# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/v1/mikan/search":{"post":{"tags":["Mikan"],"summary":"Run Engine","operationId":"run_engine_v1_mikan_search_post","parameters":[{"name":"x-token","in":"header","required":true,"schema":{"type":"string","title":"X-Token"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MikanSearchPostBody"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MikanSearchResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/user/create":{"post":{"tags":["User"],"summary":" Create User","operationId":"_create_user_v1_user_create_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InfoMessage"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/user/all":{"get":{"tags":["User"],"summary":" Get Users","operationId":"_get_users_v1_user_all_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/UserResponse"},"type":"array","title":"Response  Get Users V1 User All Get"}}}}}}},"/v1/user/login":{"post":{"tags":["User"],"summary":" User Login","operationId":"_user_login_v1_user_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InfoMessage"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/engine/run":{"get":{"tags":["Engine"],"summary":"Run Engine","operationId":"run_engine_v1_engine_run_get","parameters":[{"name":"engine_namespace","in":"query","required":true,"schema":{"type":"string","title":"Engine Namespace"}},{"name":"x-token","in":"header","required":true,"schema":{"type":"string","title":"X-Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InfoMessage"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/engine/run-all":{"get":{"tags":["Engine"],"summary":"Refresh All Engines","operationId":"refresh_all_engines_v1_engine_run_all_get","parameters":[{"name":"x-token","in":"header","required":true,"schema":{"type":"string","title":"X-Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InfoMessage"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/download/all":{"get":{"tags":["Download"],"summary":"Get Downloads","operationId":"get_downloads_v1_download_all_get","parameters":[{"name":"x-token","in":"header","required":true,"schema":{"type":"string","title":"X-Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DownloadResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/download/{download_id}":{"get":{"tags":["Download"],"summary":"Get Download","operationId":"get_download_v1_download__download_id__get","parameters":[{"name":"download_id","in":"path","required":true,"schema":{"type":"string","title":"Download Id"}},{"name":"x-token","in":"header","required":true,"schema":{"type":"string","title":"X-Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DownloadItem"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/download/pause/{download_id}":{"get":{"tags":["Download"],"summary":"Pause Download","operationId":"pause_download_v1_download_pause__download_id__get","parameters":[{"name":"download_id","in":"path","required":true,"schema":{"type":"string","title":"Download Id"}},{"name":"x-token","in":"header","required":true,"schema":{"type":"string","title":"X-Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DownloadItem"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/download/resume/{download_id}":{"get":{"tags":["Download"],"summary":"Resume Download","operationId":"resume_download_v1_download_resume__download_id__get","parameters":[{"name":"download_id","in":"path","required":true,"schema":{"type":"string","title":"Download Id"}},{"name":"x-token","in":"header","required":true,"schema":{"type":"string","title":"X-Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DownloadItem"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/download/remove/{download_id}":{"get":{"tags":["Download"],"summary":"Remove Download","operationId":"remove_download_v1_download_remove__download_id__get","parameters":[{"name":"download_id","in":"path","required":true,"schema":{"type":"string","title":"Download Id"}},{"name":"x-token","in":"header","required":true,"schema":{"type":"string","title":"X-Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DownloadItem"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/media/cover/{media_id}":{"get":{"tags":["Media"],"summary":" Get Cover Image","operationId":"_get_cover_image_v1_media_cover__media_id__get","parameters":[{"name":"media_id","in":"path","required":true,"schema":{"type":"string","title":"Media Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/media/all":{"get":{"tags":["Media"],"summary":" Get All Media","operationId":"_get_all_media_v1_media_all_get","parameters":[{"name":"x-token","in":"header","required":true,"schema":{"type":"string","title":"X-Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MediasResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/media/{media_id}":{"get":{"tags":["Media"],"summary":" Get Media","operationId":"_get_media_v1_media__media_id__get","parameters":[{"name":"media_id","in":"path","required":true,"schema":{"type":"string","title":"Media Id"}},{"name":"x-token","in":"header","required":true,"schema":{"type":"string","title":"X-Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MediaResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/subscribe/all":{"get":{"tags":["Subscribe"],"summary":"Get Downloads","operationId":"get_downloads_v1_subscribe_all_get","parameters":[{"name":"x-token","in":"header","required":true,"schema":{"type":"string","title":"X-Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubscriptionsAllResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/subscribe/add":{"post":{"tags":["Subscribe"],"summary":"Add Subscription","operationId":"add_subscription_v1_subscribe_add_post","parameters":[{"name":"x-token","in":"header","required":true,"schema":{"type":"string","title":"X-Token"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddSubscriptionBody"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InfoMessage"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/plugin/info":{"get":{"tags":["Plugin"],"summary":"Run Engine","operationId":"run_engine_v1_plugin_info_get","parameters":[{"name":"x-token","in":"header","required":true,"schema":{"type":"string","title":"X-Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PluginInfoResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AddSubscriptionBody":{"properties":{"name":{"type":"string","title":"Name"},"data":{"type":"string","title":"Data"},"namespace":{"type":"string","title":"Namespace"}},"type":"object","required":["name","data","namespace"],"title":"AddSubscriptionBody"},"Content":{"properties":{"title":{"type":"string","title":"Title"},"episode":{"type":"integer","title":"Episode"},"link":{"type":"string","title":"Link"},"path":{"type":"string","title":"Path"},"add_time":{"type":"integer","title":"Add Time"}},"type":"object","required":["title","episode","link","path","add_time"],"title":"Content"},"DownloadData":{"properties":{"id":{"type":"string","title":"Id"},"is_metadata":{"type":"boolean","title":"Is Metadata"},"name":{"type":"string","title":"Name"},"target_path":{"type":"string","title":"Target Path"},"dir":{"type":"string","title":"Dir"},"total_length":{"type":"integer","title":"Total Length"},"progress":{"type":"number","title":"Progress"},"current_download":{"type":"integer","title":"Current Download"},"status":{"type":"string","title":"Status"},"current_speed":{"type":"integer","title":"Current Speed"}},"type":"object","required":["id","is_metadata","name","target_path","dir","total_length","progress","current_download","status","current_speed"],"title":"DownloadData"},"DownloadItem":{"properties":{"message":{"type":"string","title":"Message","default":"Success"},"success":{"type":"boolean","title":"Success","default":true},"title":{"type":"string","title":"Title","default":"Info"},"data":{"anyOf":[{"$ref":"#/components/schemas/DownloadData"},{"type":"null"}]}},"type":"object","title":"DownloadItem"},"DownloadResponse":{"properties":{"message":{"type":"string","title":"Message","default":"Success"},"success":{"type":"boolean","title":"Success","default":true},"title":{"type":"string","title":"Title","default":"Info"},"data":{"anyOf":[{"items":{"$ref":"#/components/schemas/DownloadData"},"type":"array"},{"type":"null"}],"title":"Data"}},"type":"object","title":"DownloadResponse"},"Engine":{"properties":{"name":{"type":"string","title":"Name"},"description":{"type":"string","title":"Description"},"namespace":{"type":"string","title":"Namespace"},"cron_str":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cron Str"}},"type":"object","required":["name","description","namespace","cron_str"],"title":"Engine"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"InfoMessage":{"properties":{"message":{"type":"string","title":"Message","default":"Success"},"success":{"type":"boolean","title":"Success","default":true},"title":{"type":"string","title":"Title","default":"Info"},"data":{"type":"string","title":"Data","default":""}},"type":"object","title":"InfoMessage"},"Media":{"properties":{"contents":{"items":{"$ref":"#/components/schemas/Content"},"type":"array","title":"Contents"},"id":{"type":"string","title":"Id"},"link":{"type":"string","title":"Link"},"title":{"type":"string","title":"Title"},"bangumi_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Bangumi Id"},"season":{"type":"integer","title":"Season"}},"type":"object","required":["contents","id","link","title","bangumi_id","season"],"title":"Media"},"MediaResponse":{"properties":{"message":{"type":"string","title":"Message","default":"Success"},"success":{"type":"boolean","title":"Success","default":true},"title":{"type":"string","title":"Title","default":"Info"},"data":{"anyOf":[{"$ref":"#/components/schemas/Media"},{"type":"null"}]}},"type":"object","title":"MediaResponse"},"MediasResponse":{"properties":{"message":{"type":"string","title":"Message","default":"Success"},"success":{"type":"boolean","title":"Success","default":true},"title":{"type":"string","title":"Title","default":"Info"},"data":{"anyOf":[{"items":{"$ref":"#/components/schemas/Media"},"type":"array"},{"type":"null"}],"title":"Data"}},"type":"object","title":"MediasResponse"},"MikanSearchPostBody":{"properties":{"keyword":{"type":"string","title":"Keyword"}},"type":"object","required":["keyword"],"title":"MikanSearchPostBody"},"MikanSearchResponse":{"properties":{"message":{"type":"string","title":"Message","default":"Success"},"success":{"type":"boolean","title":"Success","default":true},"title":{"type":"string","title":"Title","default":"Info"},"data":{"items":{"$ref":"#/components/schemas/SearchItem"},"type":"array","title":"Data","default":[]}},"type":"object","title":"MikanSearchResponse"},"PluginInfo":{"properties":{"name":{"type":"string","title":"Name"},"namespace":{"type":"string","title":"Namespace"},"description":{"type":"string","title":"Description"},"is_local_plugin":{"type":"boolean","title":"Is Local Plugin"},"is_internal":{"type":"boolean","title":"Is Internal"},"engines":{"items":{"$ref":"#/components/schemas/Engine"},"type":"array","title":"Engines"}},"type":"object","required":["name","namespace","description","is_local_plugin","is_internal","engines"],"title":"PluginInfo"},"PluginInfoResponse":{"properties":{"message":{"type":"string","title":"Message","default":"Success"},"success":{"type":"boolean","title":"Success","default":true},"title":{"type":"string","title":"Title","default":"Info"},"data":{"items":{"$ref":"#/components/schemas/PluginInfo"},"type":"array","title":"Data","default":[]}},"type":"object","title":"PluginInfoResponse"},"SearchItem":{"properties":{"title":{"type":"string","title":"Title"},"cover":{"type":"string","title":"Cover"},"bangumi_id":{"type":"integer","title":"Bangumi Id"},"bangumi_name":{"type":"string","title":"Bangumi Name"},"subtitle_group_id":{"type":"integer","title":"Subtitle Group Id"},"subtitle_group_name":{"type":"string","title":"Subtitle Group Name"},"group_type":{"type":"string","title":"Group Type"},"tags":{"items":{"type":"string"},"type":"array","title":"Tags","default":[]},"link":{"type":"string","title":"Link"},"last_updated":{"type":"string","format":"date-time","title":"Last Updated"},"is_checked":{"type":"boolean","title":"Is Checked","default":false}},"type":"object","required":["title","cover","bangumi_id","bangumi_name","subtitle_group_id","subtitle_group_name","group_type","link","last_updated"],"title":"SearchItem"},"Subscription":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"data":{"type":"string","title":"Data"}},"type":"object","required":["id","name","data"],"title":"Subscription"},"SubscriptionRecord":{"properties":{"namespace":{"type":"string","title":"Namespace"},"subscriptions":{"items":{"$ref":"#/components/schemas/Subscription"},"type":"array","title":"Subscriptions"}},"type":"object","required":["namespace","subscriptions"],"title":"SubscriptionRecord"},"SubscriptionsAllResponse":{"properties":{"message":{"type":"string","title":"Message","default":"Success"},"success":{"type":"boolean","title":"Success","default":true},"title":{"type":"string","title":"Title","default":"Info"},"data":{"anyOf":[{"items":{"$ref":"#/components/schemas/SubscriptionRecord"},"type":"array"},{"type":"null"}],"title":"Data"}},"type":"object","title":"SubscriptionsAllResponse"},"UserCreate":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"UserCreate"},"UserResponse":{"properties":{"username":{"type":"string","title":"Username"},"is_superuser":{"type":"boolean","title":"Is Superuser"}},"type":"object","required":["username","is_superuser"],"title":"UserResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}